requirements.txt
anyio==4.6.2.post1
APScheduler==3.10.4
blinker==1.4
certifi==2024.8.30
charset-normalizer==3.4.0
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyTelegramBotAPI==4.23.0
python-apt==2.4.0+ubuntu1
python-telegram-bot==21.6
pytz==2024.2
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
systemd-python==234
typing_extensions==4.12.2
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
zipp==1.0.0
